Uses go:embed for wasm stats and helm chart
The way we populate the value for wasm stats and helm charts
is by using the linker -X flag. This makes the linking command
to be extremely long (40k+ chars). This is above the ARG_MAX limit posed by cygwin.

Fixes #3573

This patch updates to go version 1.16 and uses go:embed to embed the files into variables

Co-authored-by: Jon Huhn johuhn@microsoft.com

Signed-off-by: Sotiris Nanopoulos <sonanopo@microsoft.com>